+++ src/UI/Components/DiffWizard/ItemList/Entry/index.ts
@@ -1,0 +1,1 @@
+export * from "./Entry";
