+++ src/UI/Components/DiffWizard/ItemList/Entry/utils.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const Title = styled.span`
+  line-height: 29px;
+  font-size: 1rem;
+  padding: 0 4px;
+  display: inline-block;
+`;
